Bugfixes on the parameter change BLoC.

+++ BLEApp/ble_app/lib/src/persistence/entities/deviceParameters.dart
@@ -71,7 +71,7 @@ class DeviceParameters extends Model {
     double maxCutoffDischargeCurrent,
     int maxCurrentTimeLimitPeriod,
     double maxCutoffChargeCurrent,
-    int motoHoursCounterCurrentThreshold,
+    double motoHoursCounterCurrentThreshold,
     int currentCutOffTimerPeriod,
     int maxCutoffTemperature,
     int maxTemperatureRecovery,
